requirements.txt
playwright==1.56.0
pytest==9.0.1
pytest-html==4.1.1
pyyaml==6.0.3
loguru==0.7.3
pytest-xdist==3.8.0
